MODIFY LWM2M Handlers to extract every handler to its own file
