+++ lib/iotAgentLwm2m.js
@@ -25,8 +25,9 @@
 var iotAgentLib = require('iotagent-node-lib'),
     lwm2mLib = require('lwm2m-node-lib').server,
     ngsiHandlers = require('./ngsiHandlers'),
-    lwm2mHandlers = require('./lwm2mHandlers'),
     registrationHandler = require('./services/lwm2mHandlers/registration'),
+    updateRegistrationHandler = require('./services/lwm2mHandlers/updateRegistration'),
+    unregistrationHandler = require('./services/lwm2mHandlers/unregistration'),
     commons = require('./commons'),
     logger = require('logops'),
     async = require('async'),
@@ -46,8 +47,8 @@ function initialize(callback) {
     iotAgentLib.setConfigurationHandler(ngsiHandlers.updateConfiguration);
 
     lwm2mLib.setHandler(serverInfo, 'registration', registrationHandler.handler);
-    lwm2mLib.setHandler(serverInfo, 'unregistration', lwm2mHandlers.unregistration);
-    lwm2mLib.setHandler(serverInfo, 'updateRegistration', lwm2mHandlers.updateRegistration);
+    lwm2mLib.setHandler(serverInfo, 'unregistration', unregistrationHandler.handler);
+    lwm2mLib.setHandler(serverInfo, 'updateRegistration', updateRegistrationHandler.handler);
 
     logger.info(context, 'Agent started');
     callback();
